nú skal tað rigga

+++ TestProject/src/Interfaces/IPythagoras.java
@@ -2,6 +2,6 @@ package Interfaces;
 
 public interface IPythagoras {
 	
-	double calcHypotenuse(float a, float b);
+	double calcHypotenuse(double a, double b);
 
 }
